# backend/docs/swagger.yaml
basePath: /api
definitions:
  models.ParkingLot:
    properties:
      capacity:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      parkingLotID:
        type: integer
      parkingName:
        type: string
      rates:
        type: number
    type: object
...
host: localhost:8000
info:
  contact:
    email: 1351688749@qq.com
    name: API Support
    url: http://example.com/support
  description: API documentation for Parking Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Parking Management API
  version: "1.0"
paths:
...
  /parkinglot:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all parking lots
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ParkingLot'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all parking lots
      tags:
      - ParkingLot
    post:
      consumes:
      - application/json
      description: Create a new parking lot with given details
      parameters:
      - description: Parking Lot Details
        in: body
        name: parkingLot
        required: true
        schema:
          $ref: '#/definitions/models.ParkingLot'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a parking lot
      tags:
      - ParkingLot
  /parkinglot/allincome/all:
    get:
      consumes:
      - application/json
      description: Retrieve the total income of all parking lots
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all income by lot ID
      tags:
      - ParkingLot
  /parkinglot/id/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a parking lot by its ID
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete parking lot
      tags:
      - ParkingLot
    get:
      consumes:
      - application/json
      description: Retrieve details of a parking lot by its ID
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ParkingLot'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking lot by ID
      tags:
      - ParkingLot
    put:
      consumes:
      - application/json
      description: Update the details of a parking lot
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Parking Lot Details
        in: body
        name: parkingLot
        required: true
        schema:
          $ref: '#/definitions/models.ParkingLot'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update parking lot
      tags:
      - ParkingLot
  /parkinglot/id/{id}/start/{start}/end/{end}:
    get:
      consumes:
      - application/json
      description: Retrieve the occupancy of a parking lot by its ID and time range
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start time (YYYY-MM-DDTHH:MM:SS)
        in: query
        name: start
        required: true
        type: string
      - description: End time (YYYY-MM-DDTHH:MM:SS)
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking lot occupancy by time
      tags:
      - ParkingLot
  /parkinglot/income/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the total income of a parking lot by its ID
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all income by lot ID
      tags:
      - ParkingLot
  /parkinglot/occupancy-rate/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the occupancy rate of a parking lot by its ID
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking lot occupancy rate
      tags:
      - ParkingLot
  /parkinglot/status/lot/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the status of a parking lot by its ID
      parameters:
      - description: Parking Lot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parking lot status by ID
      tags:
      - ParkingLot
swagger: "2.0"